requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
clearbit==0.1.7
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.12.0
idna==3.3
password==0.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
